Refactored tag and topic services.

+++ tag-service/src/main/java/com/gmail/javacoded78/service/impl/TagClientServiceImpl.java
@@ -1,36 +1,68 @@
 package com.gmail.javacoded78.service.impl;
 
-import com.gmail.javacoded78.common.models.Tag;
+import com.gmail.javacoded78.model.Tag;
+import com.gmail.javacoded78.model.TweetTag;
 import com.gmail.javacoded78.repository.TagRepository;
+import com.gmail.javacoded78.repository.TweetTagRepository;
 import com.gmail.javacoded78.service.TagClientService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 @Service
 @RequiredArgsConstructor
 public class TagClientServiceImpl implements TagClientService {
 
     private final TagRepository tagRepository;
+    private final TweetTagRepository tweetTagRepository;
 
     @Override
-    public List<Tag> getTagsByTweetId(Long tweetId) {
-        return tagRepository.findByTweets_Id(tweetId);
+    @Transactional
+    public void parseHashtagsInText(String text, Long tweetId) {
+        Pattern pattern = Pattern.compile("(#\\w+)\\b");
+        Matcher match = pattern.matcher(text);
+        List<String> hashtags = new ArrayList<>();
+
+        while (match.find()) {
+            hashtags.add(match.group(1));
+        }
+
+        if (!hashtags.isEmpty()) {
+            hashtags.forEach(hashtag -> {
+                Optional<Tag> tag = tagRepository.findByTagName(hashtag);
+                TweetTag tweetTag;
+
+                if (tag.isPresent()) {
+                    tweetTag = new TweetTag(tag.get().getId(), tweetId);
+                    tagRepository.updateTagQuantity(tag.get().getId(), true);
+                } else {
+                    Tag newTag = new Tag(hashtag, 1L);
+                    tagRepository.save(newTag);
+                    tweetTag = new TweetTag(newTag.getId(), tweetId);
+                }
+                tweetTagRepository.save(tweetTag);
+            });
+        }
     }
 
     @Override
-    public Tag getTagByTagName(String tagName) {
-        return tagRepository.findByTagName(tagName);
-    }
-
-    @Override
-    public Tag saveTag(Tag tag) {
-        return tagRepository.save(tag);
-    }
-
-    @Override
-    public void deleteTag(Tag tag) {
-        tagRepository.delete(tag);
+    @Transactional
+    public void deleteTagsByTweetId(Long tweetId) {
+        List<Long> tagsIds = tweetTagRepository.getTagIdsByTweetId(tweetId);
+        List<Tag> tags = tagRepository.getTagsBuIds(tagsIds);
+        tags.forEach(tag -> {
+            if (tag.getTweetsQuantity() - 1 == 0) {
+                tagRepository.delete(tag);
+                tweetTagRepository.deleteTag(tag.getId());
+            } else {
+                tagRepository.updateTagQuantity(tag.getId(), false);
+            }
+        });
     }
 }
